refactor: using getValidClassNames and seperate each style in component bb-9
