+++ apps/frontend/src/assets/css/styles.css
@@ -1,2 +1,2 @@
-@import url("./variables.css");
 @import url("./scaffolding.css");
+@import url("./variables.css");
