+++ apps/frontend/src/assets/css/variables.css
@@ -9,7 +9,6 @@
 	--light-gray: #e0e0e0;
 	--white: #ffffff;
 	--disable: #e0e0e0;
-	--nunito-font: "Nunito", sans-serif;
 	--pink-gradient: linear-gradient(
 		325deg,
 		hsl(338deg 100% 67%) 0%,
