--- apps/frontend/src/libs/components/css/button.module.css
@@ -1,18 +1,0 @@
-.btn {
-	width: 100%;
-	padding: 10px 0;
-	font-family: var(--nunito-font);
-	font-size: 16px;
-	font-weight: 700;
-	color: var(--white);
-	border: none;
-	border-radius: 32px;
-}
-
-.btn-dark {
-	background-color: var(--black);
-}
-
-.btn:disabled {
-	background-color: var(--disable);
-}
