--- apps/frontend/src/libs/components/css/input.module.css
@@ -1,39 +1,0 @@
-.input-container {
-	display: flex;
-	flex-direction: column;
-	font-family: var(--nunito-font);
-}
-
-.input-container .input-content {
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-	padding: 0 8px;
-	border: 1px solid #e0e0e0;
-	border-radius: 8px;
-}
-
-.input-container .label {
-	margin-bottom: 8px;
-	font-family: inherit;
-	font-size: 16px;
-	font-weight: 600;
-}
-
-.input-container .input {
-	width: 80%;
-	padding: 8px;
-	font-family: inherit;
-	border: none;
-}
-
-.input-container .btn-icon {
-	z-index: 2;
-	background-color: transparent;
-	border: none;
-	opacity: 0.3;
-}
-
-.input-container .input:focus {
-	outline: none;
-}
