--- apps/frontend/src/libs/components/css/link.module.css
@@ -1,7 +1,0 @@
-.link {
-	text-decoration: none;
-}
-
-.link-default {
-	color: #00c2fc;
-}
