+++ apps/frontend/src/libs/components/link/libs/types/link.type.ts
@@ -1,0 +1,3 @@
+type LinkType = "button" | "navLink";
+
+export { type LinkType };
