+++ apps/frontend/src/libs/components/link/libs/types/types.ts
@@ -1,0 +1,1 @@
+export { type LinkType } from "./link.type.js";
