+++ apps/frontend/src/libs/components/link/link.tsx
@@ -1,29 +1,24 @@
 import { NavLink } from "react-router-dom";
 
 import { type AppRoute } from "~/libs/enums/enums.js";
+import { getValidClassNames } from "~/libs/helpers/helpers.js";
 import { type ValueOf } from "~/libs/types/types.js";
 
-import styles from "../css/link.module.css";
+import { LinkType } from "./libs/types/types.js";
+import styles from "./styles.module.css";
 
 type Properties = {
 	children: React.ReactNode;
 	to: ValueOf<typeof AppRoute>;
-	variant: "default";
+	type?: LinkType;
 };
 
-const Link: React.FC<Properties> = ({
-	children,
-	to,
-	variant = "default",
-}: Properties) => {
-	const linkClass = "link-" + variant;
-	return (
-		<NavLink
-			className={`${styles["link"] as string} ${styles[linkClass] as string}`}
-			to={to}
-		>
-			{children}
-		</NavLink>
-	);
-};
+const Link: React.FC<Properties> = ({ children, to, type }: Properties) => (
+	<NavLink
+		className={getValidClassNames(styles["link"], type && styles[type])}
+		to={to}
+	>
+		{children}
+	</NavLink>
+);
 export { Link };
