+++ apps/frontend/src/libs/components/link/styles.module.css
@@ -1,4 +1,5 @@
 .link {
+	color: #00c2fc;
 	text-decoration: none;
 }
