+++ apps/frontend/src/pages/auth/components/sign-in-form/sign-in-form.tsx
@@ -41,9 +41,7 @@ const SignInForm: React.FC<Properties> = ({ onSubmit }: Properties) => {
 						<h1 className={styles["form-header__text"]}>SIGN IN</h1>
 						<span className={styles["form-header__sub-text"]}>
 							No account? Go to{" "}
-							<Link to={AppRoute.SIGN_UP} variant="default">
-								Create an Account
-							</Link>
+							<Link to={AppRoute.SIGN_UP}>Create an Account</Link>
 						</span>
 					</div>
