+++ packages/shared/src/modules/users/libs/types/user-sign-up-request-dto.type.ts
@@ -1,5 +1,4 @@
 type UserSignUpRequestDto = {
-	confirmPassword?: string;
 	email: string;
 	name: string;
 	password: string;
